+++ features/navigation/controls/NavigationActionsController.tsx
@@ -20,6 +20,7 @@ export function NavigationActionsController({ isConnected }: NavigationActionsCo
   const isViewBelowM = useMediaQuery(`(max-width: ${navigationBreakpoints[1] - 1}px)`)
 
   const { UseNetworkSwitcher: useNetworkSwitcher } = getAppConfig('features')
+  console.log('useNetworkSwitcher', useNetworkSwitcher)
 
   return (
     <>
