+++ helpers/config/constants.ts
@@ -1,2 +1,2 @@
-export const configCacheTime = 60 * 1 // 5 minutes
+export const configCacheTime = 5 * 2 // 2 minutes
 export const configLSKey = 'ob-config'
