+++ helpers/config/index.ts
@@ -6,5 +6,4 @@ export type {
   ConfigContext,
 } from './types'
 export { getAppConfig, getAppConfigSync, saveConfigToLocalStorage } from './access-config-context'
-export { setupConfigContext, emptyConfig } from './setup-config-context'
 export { configCacheTime, configLSKey } from './constants'
